Refactored package into src structure
