+++ .gitignore
@@ -1,1 +1,3 @@
 __pycache__/
+decombinator.egg-info
+.pytest_cache
