+++ pyproject.toml
@@ -1,0 +1,16 @@
+[project]
+name = "decombinator"
+version = "2024.0.0a1"
+description = "Fast, error-correcting analysis of TCR repertoires"
+
+[build-system]
+requires = ["setuptools >= 61.0.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.pytest.ini_options]
+pythonpath = [
+  "./src"
+]
