+++ requirements.txt
@@ -1,0 +1,8 @@
+biopython==1.83
+-e git+https://github.com/innate2adaptive/Decombinator.git@80392f4e89b9b61aae94400dc8a618c0fd564e4e#egg=decombinator
+numpy==1.26.4
+pandas==2.2.2
+python-dateutil==2.9.0.post0
+pytz==2024.1
+six==1.16.0
+tzdata==2024.1
